backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic==1.10.13
email-validator==2.2.0
python-multipart==0.0.19

